apiservice/requirements.txt
appdirs==1.4.0
appnope==0.1.0
attrs==16.3.0
backports.shutil-get-terminal-size==1.0.0
click==6.7
coreapi==2.2.4
coreschema==0.0.4
decorator==4.0.11
diff-match-patch==20121119
Django==1.10.5
django-cors-headers==2.0.2
django-debug-toolbar==1.6
django-extensions==1.7.6
django-filter==1.0.1
django-import-export==0.5.1
django-rest-swagger==2.1.1
djangorestframework==3.5.4
docutils==0.13.1
enum34==1.1.6
ipython==5.2.2
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
jsonref==0.1
markdown2==2.3.3
MarkupSafe==0.23
mysqlclient==1.3.10
nose==1.3.7
openapi-codec==1.3.1
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psutil==5.1.3
ptyprocess==0.5.1
Pygments==2.2.0
Pympler==0.4.3
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
requests==2.13.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
tablib==0.11.4
termcolor==1.1.0
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.11.15
whitenoise==3.3.0
xmltodict==0.10.2
